Backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==35.0.0
Django==3.2.8
djangorestframework==3.12.4
docopt==0.6.2
idna==3.2
mysqlclient==2.0.3
numpy==1.21.4
pandas==1.3.4
pycparser==2.20
PyJWT==2.2.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
yarg==0.1.9
